Dedupe Photos-asset URL parsing into PhotosAssetURL
The same photos://asset?id=... parse-then-fetch-PHAsset snippet was
copy-pasted in 6 places (UniversalImageView, QuickLookCoordinator x2,
AppViewModel, PhotoLibraryScanner, FileDeleter). Centralize it in
Core/PhotosAssetURL.swift (build/localIdentifier/asset) and have every
call site use it instead. Pure refactor, no behavior change.
